refactor(storage): Split large files and fix critical encoding bugs
Bug fixes:
- Fix negative timestamp nanosecond calculation for pre-epoch dates
- Return error on timestamp overflow instead of silently using 0
- Return error for unsupported types instead of silently encoding as NULL
- Improve error messages with index name, doc_id, and marker byte info

Test improvements:
- Add decode error path tests (truncated buffers, malformed escapes)
- Add UniqueIndex update constraint violation test
- Add Float NaN sort order verification
- Add comprehensive Float32 tests
- Add Nillable variant tests

File structure refactoring:
- Split encoding.rs (~1700 lines) into focused modules:
  null.rs, bool.rs, varint.rs, float.rs, bytes.rs, time.rs, fixed.rs
- Split index.rs (~940 lines) into focused modules:
  traits.rs, simple.rs, unique.rs

Also includes formatting fixes for crypto and db crates.
