+++ crates/storage/src/field_value.rs
@@ -24,127 +24,108 @@ use crate::encoding::{self, EncodedType};
 pub use crate::keys::datastore::IndexedField;
 
 /// Encode a NormalValue to bytes using order-preserving encoding.
-///
-/// This matches Go's `EncodeFieldValue` in encoding/field_value.go
-pub fn encode_field_value(buf: Vec<u8>, val: &NormalValue, descending: bool) -> Vec<u8> {
+pub fn encode_field_value(buf: Vec<u8>, val: &NormalValue, descending: bool) -> Result<Vec<u8>> {
     if val.is_nil() {
-        return if descending {
+        return Ok(if descending {
             encoding::encode_null_descending(buf)
         } else {
             encoding::encode_null_ascending(buf)
-        };
+        });
     }
 
     match val {
-        NormalValue::Bool(v) => {
-            if descending {
-                encoding::encode_bool_descending(buf, *v)
-            } else {
-                encoding::encode_bool_ascending(buf, *v)
-            }
-        }
-        NormalValue::NillableBool(Some(v)) => {
-            if descending {
-                encoding::encode_bool_descending(buf, *v)
-            } else {
-                encoding::encode_bool_ascending(buf, *v)
-            }
-        }
-        NormalValue::Int(v) => {
-            if descending {
-                encoding::encode_varint_descending(buf, *v)
-            } else {
-                encoding::encode_varint_ascending(buf, *v)
-            }
-        }
-        NormalValue::NillableInt(Some(v)) => {
-            if descending {
-                encoding::encode_varint_descending(buf, *v)
-            } else {
-                encoding::encode_varint_ascending(buf, *v)
-            }
-        }
-        NormalValue::Float32(v) => {
-            if descending {
-                encoding::encode_float32_descending(buf, *v)
-            } else {
-                encoding::encode_float32_ascending(buf, *v)
-            }
-        }
-        NormalValue::NillableFloat32(Some(v)) => {
-            if descending {
-                encoding::encode_float32_descending(buf, *v)
-            } else {
-                encoding::encode_float32_ascending(buf, *v)
-            }
-        }
-        NormalValue::Float64(v) => {
-            if descending {
-                encoding::encode_float64_descending(buf, *v)
-            } else {
-                encoding::encode_float64_ascending(buf, *v)
-            }
-        }
-        NormalValue::NillableFloat64(Some(v)) => {
-            if descending {
-                encoding::encode_float64_descending(buf, *v)
-            } else {
-                encoding::encode_float64_ascending(buf, *v)
-            }
-        }
-        NormalValue::String(v) => {
-            if descending {
-                encoding::encode_string_descending(buf, v)
-            } else {
-                encoding::encode_string_ascending(buf, v)
-            }
-        }
-        NormalValue::NillableString(Some(v)) => {
-            if descending {
-                encoding::encode_string_descending(buf, v)
-            } else {
-                encoding::encode_string_ascending(buf, v)
-            }
-        }
-        NormalValue::Bytes(v) => {
-            if descending {
-                encoding::encode_bytes_descending(buf, v)
-            } else {
-                encoding::encode_bytes_ascending(buf, v)
-            }
-        }
-        NormalValue::NillableBytes(Some(v)) => {
-            if descending {
-                encoding::encode_bytes_descending(buf, v)
-            } else {
-                encoding::encode_bytes_ascending(buf, v)
-            }
-        }
+        NormalValue::Bool(v) => Ok(if descending {
+            encoding::encode_bool_descending(buf, *v)
+        } else {
+            encoding::encode_bool_ascending(buf, *v)
+        }),
+        NormalValue::NillableBool(Some(v)) => Ok(if descending {
+            encoding::encode_bool_descending(buf, *v)
+        } else {
+            encoding::encode_bool_ascending(buf, *v)
+        }),
+        NormalValue::Int(v) => Ok(if descending {
+            encoding::encode_varint_descending(buf, *v)
+        } else {
+            encoding::encode_varint_ascending(buf, *v)
+        }),
+        NormalValue::NillableInt(Some(v)) => Ok(if descending {
+            encoding::encode_varint_descending(buf, *v)
+        } else {
+            encoding::encode_varint_ascending(buf, *v)
+        }),
+        NormalValue::Float32(v) => Ok(if descending {
+            encoding::encode_float32_descending(buf, *v)
+        } else {
+            encoding::encode_float32_ascending(buf, *v)
+        }),
+        NormalValue::NillableFloat32(Some(v)) => Ok(if descending {
+            encoding::encode_float32_descending(buf, *v)
+        } else {
+            encoding::encode_float32_ascending(buf, *v)
+        }),
+        NormalValue::Float64(v) => Ok(if descending {
+            encoding::encode_float64_descending(buf, *v)
+        } else {
+            encoding::encode_float64_ascending(buf, *v)
+        }),
+        NormalValue::NillableFloat64(Some(v)) => Ok(if descending {
+            encoding::encode_float64_descending(buf, *v)
+        } else {
+            encoding::encode_float64_ascending(buf, *v)
+        }),
+        NormalValue::String(v) => Ok(if descending {
+            encoding::encode_string_descending(buf, v)
+        } else {
+            encoding::encode_string_ascending(buf, v)
+        }),
+        NormalValue::NillableString(Some(v)) => Ok(if descending {
+            encoding::encode_string_descending(buf, v)
+        } else {
+            encoding::encode_string_ascending(buf, v)
+        }),
+        NormalValue::Bytes(v) => Ok(if descending {
+            encoding::encode_bytes_descending(buf, v)
+        } else {
+            encoding::encode_bytes_ascending(buf, v)
+        }),
+        NormalValue::NillableBytes(Some(v)) => Ok(if descending {
+            encoding::encode_bytes_descending(buf, v)
+        } else {
+            encoding::encode_bytes_ascending(buf, v)
+        }),
         NormalValue::Time(v) => {
-            // Convert DateTime to nanoseconds since Unix epoch
-            let nanos = v.timestamp_nanos_opt().unwrap_or(0);
-            if descending {
+            let nanos = v.timestamp_nanos_opt().ok_or_else(|| {
+                crate::corekv::Error::Other(format!(
+                    "timestamp {} cannot be encoded as nanoseconds (overflow)",
+                    v
+                ))
+            })?;
+            Ok(if descending {
                 encoding::encode_time_descending(buf, nanos)
             } else {
                 encoding::encode_time_ascending(buf, nanos)
-            }
+            })
         }
         NormalValue::NillableTime(Some(v)) => {
-            let nanos = v.timestamp_nanos_opt().unwrap_or(0);
-            if descending {
+            let nanos = v.timestamp_nanos_opt().ok_or_else(|| {
+                crate::corekv::Error::Other(format!(
+                    "timestamp {} cannot be encoded as nanoseconds (overflow)",
+                    v
+                ))
+            })?;
+            Ok(if descending {
                 encoding::encode_time_descending(buf, nanos)
             } else {
                 encoding::encode_time_ascending(buf, nanos)
-            }
+            })
         }
-        // For types we don't support yet, encode as null
-        _ => {
-            if descending {
-                encoding::encode_null_descending(buf)
-            } else {
-                encoding::encode_null_ascending(buf)
-            }
-        }
+        // Nillable None variants are handled by is_nil() check above
+        // Unsupported types return an error
+        _ => Err(crate::corekv::Error::Other(format!(
+            "unsupported field type for indexing: {:?}",
+            val
+        ))),
     }
 }
 
@@ -204,14 +185,10 @@ pub fn decode_field_value<'a>(
                 encoding::decode_bytes_ascending(buf)?
             };
             // Determine if this should be a String based on field kind
-            let is_string = matches!(
-                kind,
-                FieldKind::Scalar(ScalarKind::String)
-            );
+            let is_string = matches!(kind, FieldKind::Scalar(ScalarKind::String));
             if is_string {
-                let s = String::from_utf8(v).map_err(|e| {
-                    crate::corekv::Error::Other(format!("invalid utf-8: {}", e))
-                })?;
+                let s = String::from_utf8(v)
+                    .map_err(|e| crate::corekv::Error::Other(format!("invalid utf-8: {}", e)))?;
                 Ok((rest, NormalValue::String(s)))
             } else {
                 Ok((rest, NormalValue::Bytes(v)))
@@ -223,23 +200,34 @@ pub fn decode_field_value<'a>(
             } else {
                 encoding::decode_time_ascending(buf)?
             };
-            // Convert nanoseconds to DateTime
-            let secs = nanos / 1_000_000_000;
-            let nsecs = (nanos % 1_000_000_000) as u32;
+            // Convert nanoseconds to DateTime, handling negative timestamps correctly
+            let (secs, nsecs) = if nanos >= 0 {
+                (nanos / 1_000_000_000, (nanos % 1_000_000_000) as u32)
+            } else {
+                // For negative nanoseconds, adjust to get positive subsecond component
+                let secs = (nanos - 999_999_999) / 1_000_000_000;
+                let nsecs = ((nanos % 1_000_000_000) + 1_000_000_000) % 1_000_000_000;
+                (secs, nsecs as u32)
+            };
             let dt = Utc.timestamp_opt(secs, nsecs).single().ok_or_else(|| {
-                crate::corekv::Error::Other("invalid timestamp".to_string())
+                crate::corekv::Error::Other(format!(
+                    "invalid timestamp: cannot construct DateTime from secs={}, nsecs={}",
+                    secs, nsecs
+                ))
             })?;
             Ok((rest, NormalValue::Time(dt)))
         }
         _ => Err(crate::corekv::Error::Other(format!(
-            "cannot decode field value: unknown type {:?}",
-            typ
+            "cannot decode field value: unknown type {:?} (marker byte: 0x{:02x}, buffer len: {})",
+            typ,
+            buf.first().unwrap_or(&0),
+            buf.len()
         ))),
     }
 }
 
 /// Encode an IndexedField to bytes
-pub fn encode_indexed_field(buf: Vec<u8>, field: &IndexedField) -> Vec<u8> {
+pub fn encode_indexed_field(buf: Vec<u8>, field: &IndexedField) -> Result<Vec<u8>> {
     encode_field_value(buf, &field.value, field.descending)
 }
 
@@ -254,12 +242,12 @@ mod tests {
 
         for v in test_cases {
             let val = NormalValue::Bool(v);
-            let buf = encode_field_value(vec![], &val, false);
+            let buf = encode_field_value(vec![], &val, false).unwrap();
             let (_, decoded) = decode_field_value(&buf, false, &FieldKind::bool()).unwrap();
             assert_eq!(decoded, val);
 
             // Test descending
-            let buf = encode_field_value(vec![], &val, true);
+            let buf = encode_field_value(vec![], &val, true).unwrap();
             let (_, decoded) = decode_field_value(&buf, true, &FieldKind::bool()).unwrap();
             assert_eq!(decoded, val);
         }
@@ -271,12 +259,12 @@ mod tests {
 
         for v in test_cases {
             let val = NormalValue::Int(v);
-            let buf = encode_field_value(vec![], &val, false);
+            let buf = encode_field_value(vec![], &val, false).unwrap();
             let (_, decoded) = decode_field_value(&buf, false, &FieldKind::int()).unwrap();
             assert_eq!(decoded, val);
 
             // Test descending
-            let buf = encode_field_value(vec![], &val, true);
+            let buf = encode_field_value(vec![], &val, true).unwrap();
             let (_, decoded) = decode_field_value(&buf, true, &FieldKind::int()).unwrap();
             assert_eq!(decoded, val);
         }
@@ -288,7 +276,7 @@ mod tests {
 
         for v in test_cases {
             let val = NormalValue::Float64(v);
-            let buf = encode_field_value(vec![], &val, false);
+            let buf = encode_field_value(vec![], &val, false).unwrap();
             let (_, decoded) = decode_field_value(&buf, false, &FieldKind::float64()).unwrap();
             assert_eq!(decoded, val);
         }
@@ -300,12 +288,12 @@ mod tests {
 
         for v in test_cases {
             let val = NormalValue::String(v.to_string());
-            let buf = encode_field_value(vec![], &val, false);
+            let buf = encode_field_value(vec![], &val, false).unwrap();
             let (_, decoded) = decode_field_value(&buf, false, &FieldKind::string()).unwrap();
             assert_eq!(decoded, val);
 
             // Test descending
-            let buf = encode_field_value(vec![], &val, true);
+            let buf = encode_field_value(vec![], &val, true).unwrap();
             let (_, decoded) = decode_field_value(&buf, true, &FieldKind::string()).unwrap();
             assert_eq!(decoded, val);
         }
@@ -316,7 +304,7 @@ mod tests {
         let dt = Utc.with_ymd_and_hms(2024, 1, 15, 12, 30, 45).unwrap();
         let val = NormalValue::Time(dt);
 
-        let buf = encode_field_value(vec![], &val, false);
+        let buf = encode_field_value(vec![], &val, false).unwrap();
         let (_, decoded) = decode_field_value(&buf, false, &FieldKind::datetime()).unwrap();
         assert_eq!(decoded.as_time(), val.as_time());
     }
@@ -325,12 +313,12 @@ mod tests {
     fn test_encode_decode_null() {
         let val = NormalValue::Null;
 
-        let buf = encode_field_value(vec![], &val, false);
+        let buf = encode_field_value(vec![], &val, false).unwrap();
         let (_, decoded) = decode_field_value(&buf, false, &FieldKind::int()).unwrap();
         assert!(decoded.is_nil());
 
         // Test descending null
-        let buf = encode_field_value(vec![], &val, true);
+        let buf = encode_field_value(vec![], &val, true).unwrap();
         let (_, decoded) = decode_field_value(&buf, true, &FieldKind::int()).unwrap();
         assert!(decoded.is_nil());
     }
@@ -340,7 +328,7 @@ mod tests {
         let values: Vec<i64> = vec![-100, -1, 0, 1, 100];
         let encoded: Vec<Vec<u8>> = values
             .iter()
-            .map(|v| encode_field_value(vec![], &NormalValue::Int(*v), false))
+            .map(|v| encode_field_value(vec![], &NormalValue::Int(*v), false).unwrap())
             .collect();
 
         for i in 0..encoded.len() - 1 {
@@ -358,7 +346,7 @@ mod tests {
         let values: Vec<i64> = vec![-100, -1, 0, 1, 100];
         let encoded: Vec<Vec<u8>> = values
             .iter()
-            .map(|v| encode_field_value(vec![], &NormalValue::Int(*v), true))
+            .map(|v| encode_field_value(vec![], &NormalValue::Int(*v), true).unwrap())
             .collect();
 
         // In descending, larger values should have smaller byte sequences
@@ -377,7 +365,9 @@ mod tests {
         let values = vec!["", "a", "aa", "ab", "b", "ba"];
         let encoded: Vec<Vec<u8>> = values
             .iter()
-            .map(|v| encode_field_value(vec![], &NormalValue::String(v.to_string()), false))
+            .map(|v| {
+                encode_field_value(vec![], &NormalValue::String(v.to_string()), false).unwrap()
+            })
             .collect();
 
         for i in 0..encoded.len() - 1 {
@@ -393,13 +383,81 @@ mod tests {
     #[test]
     fn test_indexed_field() {
         let field = IndexedField::ascending(NormalValue::Int(42));
-        let buf = encode_indexed_field(vec![], &field);
+        let buf = encode_indexed_field(vec![], &field).unwrap();
         let (_, decoded) = decode_field_value(&buf, false, &FieldKind::int()).unwrap();
         assert_eq!(decoded, NormalValue::Int(42));
 
         let field = IndexedField::descending(NormalValue::Int(42));
-        let buf = encode_indexed_field(vec![], &field);
+        let buf = encode_indexed_field(vec![], &field).unwrap();
         let (_, decoded) = decode_field_value(&buf, true, &FieldKind::int()).unwrap();
         assert_eq!(decoded, NormalValue::Int(42));
+    }
+
+    #[test]
+    fn test_pre_epoch_timestamp() {
+        // Test a date before Unix epoch (1969-06-15)
+        let dt = Utc.with_ymd_and_hms(1969, 6, 15, 12, 30, 45).unwrap();
+        let val = NormalValue::Time(dt);
+
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::datetime()).unwrap();
+        assert_eq!(decoded.as_time(), val.as_time());
+
+        // Test descending
+        let buf = encode_field_value(vec![], &val, true).unwrap();
+        let (_, decoded) = decode_field_value(&buf, true, &FieldKind::datetime()).unwrap();
+        assert_eq!(decoded.as_time(), val.as_time());
+    }
+
+    #[test]
+    fn test_nillable_variants() {
+        // Test NillableBool
+        let val = NormalValue::NillableBool(Some(true));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::bool()).unwrap();
+        assert_eq!(decoded, NormalValue::Bool(true));
+
+        // Test NillableInt
+        let val = NormalValue::NillableInt(Some(42));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::int()).unwrap();
+        assert_eq!(decoded, NormalValue::Int(42));
+
+        // Test NillableFloat32
+        let val = NormalValue::NillableFloat32(Some(3.14));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::float32()).unwrap();
+        assert_eq!(decoded, NormalValue::Float32(3.14));
+
+        // Test NillableFloat64
+        let val = NormalValue::NillableFloat64(Some(3.14159));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::float64()).unwrap();
+        assert_eq!(decoded, NormalValue::Float64(3.14159));
+
+        // Test NillableString
+        let val = NormalValue::NillableString(Some("hello".to_string()));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::string()).unwrap();
+        assert_eq!(decoded, NormalValue::String("hello".to_string()));
+
+        // Test NillableBytes
+        let val = NormalValue::NillableBytes(Some(vec![1, 2, 3]));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::blob()).unwrap();
+        assert_eq!(decoded, NormalValue::Bytes(vec![1, 2, 3]));
+
+        // Test NillableTime
+        let dt = Utc.with_ymd_and_hms(2024, 1, 15, 12, 30, 45).unwrap();
+        let val = NormalValue::NillableTime(Some(dt));
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::datetime()).unwrap();
+        assert_eq!(decoded.as_time(), Some(&dt));
+
+        // Test Nillable None variants encode as null
+        let val = NormalValue::NillableInt(None);
+        let buf = encode_field_value(vec![], &val, false).unwrap();
+        let (_, decoded) = decode_field_value(&buf, false, &FieldKind::int()).unwrap();
+        assert!(decoded.is_nil());
     }
 }
